+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ```
     * Fix the problem who prevents the merchant to add Nuvei Payment plan to the product in Prestasho v8.1.*.
     * Fix the problem where the client can combine ordinary product with Nuvei Rebilling product.
+    * In case the plugin cancel to add a product to the Cart, and there is error, show first error message.
     * This version of the plugin does not support the versions before Prestashop v8.1.*.
 ```